+++ lib/casting/cast.dart
@@ -5,3 +5,5 @@ export './cast_media_status.dart';
 export './cast_session.dart';
 export './cast_device.dart';
 export './cast_media.dart';
+export './cast_media_track.dart';
+export './cast_media_text_track_style.dart';